Move presentation elements to pptx namespace
In order to move elements to the proper namespaces and facilitate the
moving of the newly created elements to external libraries I moved all
pptx elements to a folder and namespace.

Amos King @adkron <amos@binarynoggin.com>
