--- lib/openxml/elements/shape_non_visual.rb
@@ -1,8 +1,0 @@
-module OpenXml
-  module Elements
-    class ShapeNonVisual < OpenXml::Container
-      namespace :p
-      tag :nvSpPr
-    end
-  end
-end
