--- lib/openxml/elements/shape_properties.rb
@@ -1,8 +1,0 @@
-module OpenXml
-  module Elements
-    class ShapeProperties < OpenXml::Container
-      namespace :p
-      tag :spPr
-    end
-  end
-end
