--- lib/openxml/elements/text_body.rb
@@ -1,8 +1,0 @@
-module OpenXml
-  module Elements
-    class TextBody < OpenXml::Container
-      namespace :p
-      tag :txBody
-    end
-  end
-end
