+++ lib/openxml/pptx/elements/color_map.rb
