+++ lib/openxml/pptx/elements/group_shape_properties.rb
