+++ lib/openxml/pptx/elements/nonvisual_drawing_properties.rb
