+++ lib/openxml/pptx/elements/nonvisual_group_shape_drawing_properties.rb
