+++ lib/openxml/pptx/elements/nonvisual_shape_drawing_properties.rb
