+++ lib/openxml/pptx/elements/notes_size.rb
