+++ lib/openxml/pptx/elements/shape_non_visual.rb
@@ -1,0 +1,10 @@
+module OpenXml
+  module Pptx
+    module Elements
+      class ShapeNonVisual < OpenXml::Container
+        namespace :p
+        tag :nvSpPr
+      end
+    end
+  end
+end
