+++ lib/openxml/pptx/elements/shape_properties.rb
@@ -1,0 +1,10 @@
+module OpenXml
+  module Pptx
+    module Elements
+      class ShapeProperties < OpenXml::Container
+        namespace :p
+        tag :spPr
+      end
+    end
+  end
+end
