+++ lib/openxml/pptx/elements/slide_layout_list.rb
