+++ lib/openxml/pptx/elements/slide_size.rb
