+++ lib/openxml/pptx/elements/text_body.rb
@@ -1,0 +1,10 @@
+module OpenXml
+  module Pptx
+    module Elements
+      class TextBody < OpenXml::Container
+        namespace :p
+        tag :txBody
+      end
+    end
+  end
+end
